fix: opening multiple current neo-tree windows
When opening a new neo-tree window with the `current` position, do not
close all other `current` windows.
This allows having multiple `current` windows side-by-side on the same
tabpage.

It was possible before with the following combination (open neo-tree in
one buffer, then split it):
:e .<CR>:vs

This opened two neo-tree windows side-by-side (when
`hijack_netrw_behavior` was set to `open_current`).

However, when trying
to it the other way around (open two buffers side by side, open neo-tree
in one of them, then switch to the other buffer and open neo-tree in
that one):
:vs<CR>:Neotree position=current<CR><C-W>p:Neotree position=current

Opening the second neo-tree window closed the first one.

With this fix, both combinations lead to the same result of having two
independent neo-tree windows side-by-side with the `current` position.