fix(fs_scan): ensure parents of all expanded nodes are also scanned
On refresh a list of expanded nodes is gathered to decide what paths to
scan. However, when an expanded node lives inside a closed parent node,
parent node's number of children will be wrong.
This results in unexpected behavior with `group_with_dirs`, so this fix
ensures that all parent nodes are scanned.